export images

+++ Algorithm/include/FractalGenerator/FractalParams.h
@@ -9,8 +9,8 @@ struct FractalParams
     size_t width;
     size_t heigth;
 
-    long x{ 0 };
-    long y{ 0 };
+    double x{ 0. };
+    double y{ 0. };
 
     double zoom_x{ 400. };
     double zoom_y{ 200. };
